+++ client/js/_author/components/assessments/question_types/question_common/audio_limit.jsx
@@ -59,15 +59,16 @@ class AudioLimit extends React.Component {
   }
 
   render() {
+    const uniqId = `audio-limit-${this.props.item.id}`
     return (
       <div className="au-c-file-upload__audio-settings is-active">
         <span>Audio record limit</span>
         <div className="au-c-input au-c-input--inline">
-          <label htmlFor="audio-limit" />
+          <label htmlFor={uniqId} />
           <div className="au-c-input__contain">
             <input
               className="au-c-text-input au-c-text-input--smaller"
-              id="audio-limit"
+              id={uniqId}
               type="text"
               maxLength="3"
               value={this.state.timeLimit}
